Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 1052ms, Memory: 0MB, Status: Time Limit Exceed - CodeTree

+++ 241115/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -31,7 +31,7 @@ int main() {
     cin >> n; long long a, b; cin >> a >> b;
     int mi = n; int ma = 0;
     for(int i=a;i<=b;i++){
-        long long res = search(i);
+        int res = search(i);
         mi = min(mi,res);
         ma = max(ma,res);
     }
